requirements.txt
torch
torchaudio
numpy
pandas
matplotlib
scikit-learn
librosa
PyYAML
soundfile
scipy
tqdm